+++ commands/root.go
@@ -7,6 +7,7 @@ import (
 
 	authv1 "github.com/ionos-cloud/ionosctl/commands/auth-v1"
 	cloudapiv6 "github.com/ionos-cloud/ionosctl/commands/cloudapi-v6"
+	"github.com/ionos-cloud/ionosctl/commands/dbaas"
 	"github.com/ionos-cloud/ionosctl/internal/config"
 	"github.com/ionos-cloud/ionosctl/internal/core"
 	"github.com/mitchellh/go-homedir"
@@ -195,6 +196,8 @@ func addCommands() {
 	rootCmd.AddCommand(cloudapiv6.TemplateCmd())
 	// Auth Command
 	rootCmd.AddCommand(authv1.TokenCmd())
+	// Add DBaaS Commands
+	rootCmd.AddCommand(dbaas.DataBaseServiceCmd())
 }
 
 const helpTemplate = `USAGE: {{if .Runnable}}