+++ internal/core/command_runner.go
@@ -6,11 +6,12 @@ import (
 	"io"
 	"os"
 
-	config2 "github.com/ionos-cloud/ionosctl/internal/config"
+	"github.com/ionos-cloud/ionosctl/internal/config"
 	"github.com/ionos-cloud/ionosctl/internal/printer"
 	"github.com/ionos-cloud/ionosctl/internal/utils/clierror"
 	authV1 "github.com/ionos-cloud/ionosctl/services/auth-v1"
 	"github.com/ionos-cloud/ionosctl/services/cloudapi-v6"
+	cloudapidbaaspgsql "github.com/ionos-cloud/ionosctl/services/dbaas-postgres"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -24,7 +25,7 @@ func NewCommand(ctx context.Context, parent *Command, info CommandBuilder) *Comm
 		Example: info.Example,
 		PreRun: func(cmd *cobra.Command, args []string) {
 			// Set Printer in sync with the Output Flag
-			noHeaders, _ := cmd.Flags().GetBool(config2.ArgNoHeaders)
+			noHeaders, _ := cmd.Flags().GetBool(config.ArgNoHeaders)
 			p := getPrinter(noHeaders)
 			// Set Command to Command Builder
 			// The cmd is passed to the PreCommandCfg
@@ -36,7 +37,7 @@ func NewCommand(ctx context.Context, parent *Command, info CommandBuilder) *Comm
 		},
 		Run: func(cmd *cobra.Command, args []string) {
 			// Set Printer in sync with the Output Flag
-			noHeaders, _ := cmd.Flags().GetBool(config2.ArgNoHeaders)
+			noHeaders, _ := cmd.Flags().GetBool(config.ArgNoHeaders)
 			p := getPrinter(noHeaders)
 			// Set Buffers
 			cmd.SetIn(os.Stdin)
@@ -111,7 +112,7 @@ func NewCommandCfg(ctx context.Context, in io.Reader, p printer.PrintService, in
 		// Define cmd Command Config function for Command
 		initCfg: func(c *CommandConfig) error {
 			// Load configuration file or Env Variables once
-			if err := config2.Load(); err != nil {
+			if err := config.Load(); err != nil {
 				return err
 			}
 			// Init Clients and Services
@@ -127,6 +128,13 @@ func NewCommandCfg(ctx context.Context, in io.Reader, p printer.PrintService, in
 				return err
 			}
 			if err = c.AuthV1Services.InitServices(authClient); err != nil {
+				return err
+			}
+			dbaasPgsqlClient, err := c.CloudApiDbaasPgsqlServices.InitClient()
+			if err != nil {
+				return err
+			}
+			if err = c.CloudApiDbaasPgsqlServices.InitServices(dbaasPgsqlClient); err != nil {
 				return err
 			}
 			return nil
@@ -162,8 +170,9 @@ type CommandConfig struct {
 	initCfg func(commandConfig *CommandConfig) error
 
 	// Services
-	CloudApiV6Services cloudapi_v6.Services
-	AuthV1Services     authV1.Services
+	CloudApiV6Services         cloudapi_v6.Services
+	AuthV1Services             authV1.Services
+	CloudApiDbaasPgsqlServices cloudapidbaaspgsql.Services
 
 	// Context
 	Context context.Context
@@ -171,7 +180,7 @@ type CommandConfig struct {
 
 func getPrinter(noHeaders bool) printer.PrintService {
 	var out io.Writer
-	if viper.GetBool(config2.ArgQuiet) {
+	if viper.GetBool(config.ArgQuiet) {
 		var execOut bytes.Buffer
 		out = &execOut
 	} else {
@@ -179,5 +188,5 @@ func getPrinter(noHeaders bool) printer.PrintService {
 	}
 	printReg, err := printer.NewPrinterRegistry(out, os.Stderr, noHeaders)
 	clierror.CheckError(err, os.Stderr)
-	return printReg[viper.GetString(config2.ArgOutput)]
+	return printReg[viper.GetString(config.ArgOutput)]
 }