+++ internal/utils/watch.go
@@ -12,12 +12,15 @@ import (
 
 // State Status
 const (
-	stateDeployingStatus = "DEPLOYING"
-	stateUpdatingStatus  = "UPDATING"
-	stateActiveStatus    = "ACTIVE"
-	stateAvailableStatus = "AVAILABLE"
-	stateReadyStatus     = "READY"
-	stateFailedStatus    = "FAILED"
+	stateDeployingStatus  = "DEPLOYING"
+	stateUpdatingStatus   = "UPDATING"
+	stateBusyStatus       = "BUSY"
+	stateActiveStatus     = "ACTIVE"
+	stateAvailableStatus  = "AVAILABLE"
+	stateReadyStatus      = "READY"
+	stateDoneStatus       = "DONE"
+	stateFailedStatus     = "FAILED"
+	stateDestroyingStatus = "DESTROYING"
 )
 
 // WatchStateProgress watches the state progress of a Resource until it completes with success: meaning ACTIVE or AVAILABLE or error.
@@ -51,7 +54,7 @@ func WatchStateProgress(ctx context.Context, c *core.CommandConfig, interrogator
 				return
 			}
 			if state == nil {
-				errChan <- errors.New("error getting state")
+				errChan <- errors.New("error getting state/status")
 				return
 			}
 
@@ -61,14 +64,14 @@ func WatchStateProgress(ctx context.Context, c *core.CommandConfig, interrogator
 			case stateDeployingStatus:
 				sendingProgress(1)
 				break
-			case stateUpdatingStatus:
+			case stateUpdatingStatus, stateBusyStatus:
 				sendingProgress(50)
 				break
-			case stateActiveStatus, stateAvailableStatus, stateReadyStatus:
+			case stateActiveStatus, stateAvailableStatus, stateReadyStatus, stateDoneStatus:
 				sendingProgress(100)
 				errChan <- nil
 				return
-			case stateFailedStatus:
+			case stateFailedStatus, stateDestroyingStatus:
 				errChan <- errors.New(failed)
 				return
 			}
@@ -123,6 +126,63 @@ func WatchRequestProgress(ctx context.Context, c *core.CommandConfig, interrogat
 				return
 			case ionoscloud.RequestStatusFailed:
 				errChan <- errors.New(fmt.Sprintf("%s %s", *status, *message))
+				return
+			}
+		}
+	}()
+	return progressChan, errChan
+}
+
+// WatchDeletionProgress watches the deletion progress of a Resource until it completes with success: returning 404 http response status code.
+func WatchDeletionProgress(ctx context.Context, c *core.CommandConfig, interrogator InterrogateDeletionFunc, resourceId string) (<-chan int, <-chan error) {
+	errChan := make(chan error, 1)
+	progressChan := make(chan int)
+	go func() {
+		defer close(errChan)
+		defer close(progressChan)
+		ticker := time.NewTicker(pollTime)
+		sendingProgress := func(p int) {
+			select {
+			case progressChan <- p:
+				break
+			default:
+				break
+			}
+		}
+		for {
+			select {
+			case <-ctx.Done():
+				errChan <- ctx.Err()
+				return
+			case <-ticker.C:
+				break
+			}
+
+			httpResponseCode, err := interrogator(c, resourceId)
+			if err != nil {
+				errChan <- err
+				return
+			}
+			if httpResponseCode == nil {
+				errChan <- errors.New("error getting http response status code")
+				return
+			}
+
+			// Check Resource Http Response Status Code
+			// Send Progress, Send Error if any
+			switch *httpResponseCode {
+			case 200:
+				sendingProgress(1)
+				break
+			case 202:
+				sendingProgress(50)
+				break
+			case 404:
+				sendingProgress(100)
+				errChan <- nil
+				return
+			case 400, 401, 403, 500:
+				errChan <- errors.New(failed)
 				return
 			}
 		}